실행 가능하게 수정
