--- CH02/Minki/week_1/about.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <body>
-    <!-- HEADER -->
-    <header>
-      <h1>안녕하세요 김민기입니다</h1>
-    </header>
-    <article>
-      <h3>여기는 ABOUT입니다.</h3>
-    </article>
-  </body>
-</html>
