--- CH02/Minki/week_1/other.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <body>
-    <!-- HEADER -->
-    <header>
-      <h1>안녕하세요 이정환입니다</h1>
-    </header>
-    <article>
-      <!-- 페이지별로 필요한 요소를 작성합니다 -->
-    </article>
-  </body>
-</html>ㄴ
