--- CH02/Minki/week_1/src copy/1.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <script>
-      function onClickButton() {
-        const h1Elm = document.getElementById("h1"); 
-        h1Elm.innerText = "반가워요!"; 
-      }
-    </script>
-  </head>
-  <body>
-    <h1 id="h1">안녕하세요!</h1>
-    <button onclick="onClickButton()">인사말 바꾸기</button> 
-  </body>
-</html>
