--- CH02/Minki/week_1/src copy/App.css
@@ -1,16 +1,0 @@
-body {
-    padding: 20px;
-  }
-  
-  .App {
-    margin: 0 auto;
-    width: 500px;
-  }
-  
-  .App > section { 
-    padding: 20px;
-    background-color: rgb(245, 245, 245);
-    border: 1px solid rgb(240, 240, 240);
-    border-radius: 5px;
-    margin-bottom: 10px;
-  }
