--- CH02/Minki/week_1/src copy/App.js
@@ -1,24 +1,0 @@
-import "./App.css";
-import { useState } from "react";
-import Controller from "./component/Controller";
-import Viewer from "./component/Viewer";
-
-function App() {
-  const [count, setCount] = useState(0);
-  const handleSetCount = (value) => {
-    setCount(count + value);
-  };
-
-  return (
-    <div className="App">
-      <h1>Simple Counter</h1>
-      <section>
-        <Viewer count={count} /> 
-      </section>
-      <section>
-        <Controller handleSetCount={handleSetCount} /> 
-      </section>
-    </div>
-  );
-}
-export default App;
