--- CH02/Minki/week_1/src copy/src/App.js
@@ -1,15 +1,0 @@
-import "./App.css";
-import Viewer from "./component/Viewer";
-
-function App() {
-  return (
-    <div className="App">
-      <h1>Simple Counter</h1> 
-      <section>
-        <Viewer /> 
-      </section>
-    </div>
-  );
-}
-
-export default App;
