--- CH02/Minki/week_1/src copy/src/component/Viewer.js
@@ -1,9 +1,0 @@
-const Viewer = () => {
-    return (
-      <div>
-        <div>현재 카운트: </div>
-        <h1>0</h1>
-      </div>
-    );
-  };
-  export default Viewer;s
