--- CH02/Minki/week_1/src copy/src/src/component/Controller.js
@@ -1,13 +1,0 @@
-const Controller = () => {
-    return (
-      <div>
-        <button>-1</button>
-        <button>-10</button>
-        <button>-100</button>
-        <button>+100</button>
-        <button>+10</button>
-        <button>+1</button>
-      </div>
-    );
-  };
-  export default Controller;
