--- CH02/Minki/week_1/src.App.js
@@ -1,19 +1,0 @@
-import "./App.css";
-import Controller from "./component/Controller"; 
-import Viewer from "./component/Viewer";
-
-function App() {
-  return (
-    <div className="App">
-      <h1>Simple Counter</h1>
-      <section>
-        <Viewer />
-      </section>
-      <section>
-        <Controller /> ②
-      </section>
-    </div>
-  );
-}
-
-export default App;
