+++ CH02/Minki/week_1/src/App.js
@@ -1,25 +1,25 @@
-import logo from './logo.svg';
-import './App.css';
-
+import "./App.css";
+import { useState } from "react";
+import Controller from "./component/Controller";
+import Viewer from "./component/Viewer";
+import MyHeader from "./component/MyHeader"
 function App() {
+  const [count, setCount] = useState(0);
+  const handleSetCount = (value) => {
+    setCount(count + value);
+  };
+
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <h1>Simple Counter</h1>
+      <MyHeader/>
+      <section>
+        <Viewer count={count} /> 
+      </section>
+      <section>
+        <Controller handleSetCount={handleSetCount} /> 
+      </section>
     </div>
   );
 }
-
 export default App;
