+++ CH02/Minki/week_1/src/component/MyHeader.js
@@ -5,3 +5,5 @@ function MyHeader() {
       </header>
     );
   }
+
+export default MyHeader;
